// src/notifications/dtos/notification.dto.ts
import { IsArray, IsMongoId } from 'class-validator';

export enum NotificationType {
  ORDER_CREATED = 'ORDER_CREATED',
  CLEANER_ASSIGNED = 'CLEANER_ASSIGNED',
  CLEANER_CHECKED_IN = 'CLEANER_CHECKED_IN',
  ORDER_COMPLETED = 'ORDER_COMPLETED',
  COMPLAINT_REPLIED = 'COMPLAINT_REPLIED',
  NEW_JOB_AVAILABLE = 'NEW_JOB_AVAILABLE',
  JOB_ASSIGNED = 'JOB_ASSIGNED',
  ORDER_CANCELLED = 'ORDER_CANCELLED',
  NEW_REVIEW_RECEIVED = 'NEW_REVIEW_RECEIVED',
  NEW_COMPLAINT = 'NEW_COMPLAINT',
  LOW_RATING_ALERT = 'LOW_RATING_ALERT',
  NEW_ORDER_CREATED = 'NEW_ORDER_CREATED',
  ACCOUNT_LOCKED = 'ACCOUNT_LOCKED',
  ACCOUNT_UNLOCKED = 'ACCOUNT_UNLOCKED',
}

export class CreateNotificationPayload {
  recipientId!: string;
  type!: NotificationType;
  title!: string;
  body!: string;
  referenceId?: string;
  referenceType?: string;
}

export class MarkNotificationsReadDto {
  @IsArray()
  @IsMongoId({ each: true })
  ids!: string[];
}